readability
Here are a few corrections to improve the readability of README

------------

I wasn't sure about "working with data of different sizes and length", I changed it to "different shapes and lengths" (since size and length are somewhat similar when removed from the context of code). But perhaps I made a wrong change. Please validate.

-------------

"If you want to use the binaries you need to run on Linux, use Python 3.8+ and have a CUDA GPU with CUDA11."

the last bit is confusing - what's a CUDA GPU? I rewrote it as:

"If you want to use the binaries you need to run on Linux, use Python 3.8+ and have a CUDA GPU with CUDA11."

but I think this is wrong, since pytorch binaries come with their own binary copy of cuda binaries - as you are saying this is a fork of pytorch. Perhaps you meant this is needed for building from source?

In any case please clarify what you meant by `CUDA GPU with CUDA11`

Thanks.